streamline hint parser
